+++ src/main/java/com/teamer/yapicooler/cooler/YapiCooler.java
@@ -1,6 +1,7 @@
 package com.teamer.yapicooler.cooler;
 
 import com.teamer.yapicooler.model.Group;
+import com.teamer.yapicooler.model.Project;
 import com.teamer.yapicooler.model.YapiUser;
 
 import java.io.IOException;
@@ -35,7 +36,22 @@ public interface YapiCooler {
      * 根据组列表取出组项目列表并备份
      *
      * @param groupList 组列表
+     * @return List yapi组项目列表
      * @throws IOException IO异常
      */
-    void backupGroupProject(List<Group> groupList) throws IOException;
+    List<Project> getGroupProject(List<Group> groupList) throws IOException;
+
+    /**
+     * 根据projectList导出并持久化项目接口文件
+     *
+     * @param nowDateTime 备份时间
+     * @param project     (
+     *                    id - 项目id
+     *                    name - 项目名
+     *                    groupId - 组id
+     *                    groupName - 组名
+     *                    )
+     * @throws IOException IO异常
+     */
+    void backup(String nowDateTime, Project project) throws IOException;
 }
